went back to original approach for now; stored algo in separate py file
