+++ src/main/python/translators/fluentnao/core.py
@@ -6,116 +6,32 @@ Created on 6 Jul 2013
 import math
 
 class FluentNaoTranslator(object):
-	def __init__(self):
-
-		# command: (desired deg, max deg, min deg)
-		self.commands_dict = {
-			"arms.left_forward": {
-				'LShoulderPitch': (0, 45, -45),
-				'LShoulderRoll': (0, 45, -90)
-			},
-			"arms.right_forward": {
-				'LShoulderPitch': (0, 45, -45),
-				'LShoulderRoll': (0, 90, -45)
-			}
-		}
-
-
 	def detect_command(self, joint_dict):
-		score = {}
 		commands = []
 
-		# commands
-		for command in self.commands_dict.keys():
-
-			# command joint dict: desired, max and min
-			cmd_joint_dict = self.commands_dict[command]
-
-			# final command tuple
-			cmd_tuple = (command, [0, 0]) 				 
-			
-
-			# joint degrees
-			range_match = True
-			for joint in cmd_joint_dict:
-
-				# naos current position (passed in)
-				deg_nao = math.degrees(joint_dict[joint])
-
-				# range tuple
-				cmd_range_tuple = cmd_joint_dict[joint]
-
-				# set: desired, max & min
-				deg_desired = cmd_range_tuple[0]
-				deg_max = cmd_range_tuple[1]
-				deg_min = cmd_range_tuple[2]
-
-				# in range?
-				print joint + ' ' + str(deg_min) + ' <= ' + str(deg_nao) + ' <= ' + str(deg_max)
-				if deg_min <= deg_nao <= deg_max:
-
-					# offset
-					offset = self.caculate_offset(deg_desired, deg_nao, joint)
-
-					# assign to cmd tuple
-					self.assign_offset(joint, cmd_tuple, offset)
-
-				# not in range?
-				else:
-					range_match = False
-
-			# use command
-			if range_match:
-				commands.append(cmd_tuple)
-
-		# return all commands needed for arms
-		return self.reduce_commands(commands)
-
-	def reduce_commands(self, commands):
-
-		reduced_commands = []
-    
-    	
-		for command_tuple in commands:
-
-			# match end of command
-			match = command_tuple[0].split('_')[-1]
-
-			# against collection
-			tuples = [c for c in commands if match in c[0]]
-			
-			if len(tuples) > 1:
-
-				# reduce
+		# check both arms
+		result = self.is_arms_forward(joint_dict)
+		if result:
+			commands.append(result)
+		else:
+			# check left
+			left = self.is_left_arm_forward(joint_dict)
+			if left:
+				commands.append(left)
+			else: 
+				# out or up?
 				pass
 
+			# check right
+			right = self.is_right_arm_forward(joint_dict)	
+			if right:
+				commands.append(right)
+			else:
+				# out or up?
+				pass
+
+		# return all commands needed for arms
 		return commands
-
-
-	def assign_offset(self, joint, cmd_tuple, offset):
-
-		# Roll?
-		if (joint.find('Roll') > -1):
-			
-			# first offset
-			cmd_tuple[1][1] = offset
-
-		else:
-
-			# second offset
-			cmd_tuple[1][0] = offset
-
-	def caculate_offset(self, deg_desired, deg_nao, joint):
-
-		# Roll?
-		if (joint.find('Roll') > -1):
-			
-			# Left?
-			if (joint.startswith('L')):
-				return deg_desired + deg_nao
-
-		# default offset calculation
-		return deg_desired - deg_nao
 
 	def is_left_arm_forward(self, joint_dict):
 
@@ -156,10 +72,20 @@ class FluentNaoTranslator(object):
 
 		# both forward?
 		if left and right:
-			# get offset (same for left or right)
-			pitch_offset = left[1][0]
-			roll_offset = left[1][1]
-			return ("arms.forward", [pitch_offset, roll_offset])
+			# compare offsets
+			l_pitch_offset = left[1][0]
+			l_roll_offset = left[1][1]
+			r_pitch_offset = right[1][0]
+			r_roll_offset = right[1][1]
+
+			# allow within 10 degrees
+			pitch_diff = abs(l_pitch_offset - r_pitch_offset)
+			roll_diff = abs(l_roll_offset - r_roll_offset)
+			if pitch_diff <= 10 and roll_diff <= 10:
+
+				# reduce
+				return ("arms.forward", [l_pitch_offset, l_roll_offset])
+
 		return None
